requirements.txt
geopy==2.2.*
iuliia==0.11.*
python-dotenv==0.19.*
python-telegram-bot==13.7
requests==2.26.*